Extract permission system service interface Add documentation Deprecate giveUserEntityPermissions(SecurityContext, List<String>)
